refactor: split the post capture modules and the enrichment pass
One spelling of the canonical permalink (post-capture.ts), parser regexes
derived from the template's markup constants (post-note-markup.ts), the post
frontmatter keys in one place (post-meta.ts), media localization as its own
step (post-media.ts), and the link legs extracted from
reconcileCaptureEnrichment into link-enrichment.ts over a shared
capture-enrichment-context.ts. Extension: permission/content-script/flow
modules, one X markup constants file, a shared storage mock for tests, a
respondWith helper in the background, and enable registers before persisting.
